+++ modules/frontend-sdk/eslint-plugin/.eslintrc.js
@@ -6,7 +6,12 @@
 const path = require('path');
 
 module.exports = {
+	env: {
+		node: true,
+	},
 	rules: {
+		'@liferay/no-extraneous-dependencies': 'off',
+		'@liferay/portal/no-localhost-reference': 'off',
 		'notice/notice': [
 			'error',
 			{