+++ modules/frontend-sdk/prettier-plugin/parsers.mjs
@@ -1,6 +1,6 @@
 /**
- * SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
- * SPDX-License-Identifier: MIT
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 import {format} from 'prettier';
@@ -52,8 +52,11 @@ function transformParser(parserName, defaultParser) {
 					body: formattedText,
 					type: 'FormattedText',
 				};
-			} catch (err) {
-				console.log('ERROR', options.filepath, err);
+			}
+			catch (error) {
+
+				// eslint-disable-next-line no-console
+				console.log('ERROR', options.filepath, error);
 			}
 		},
 	};