+++ modules/frontend-sdk/prettier-plugin/rules/newline-before-block-statements.mjs
@@ -1,6 +1,6 @@
 /**
- * SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
- * SPDX-License-Identifier: MIT
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
 import visitNode from '../utils/visitNode.mjs';
@@ -12,6 +12,7 @@ export function newlineBeforeBlockStatement(formattedText, ast) {
 
 	visitNode(ast, (node) => {
 		if (node.type === 'IfStatement') {
+
 			/*
 			 *                } else if {
 			 *     consequent ^         ^ alternate
@@ -37,6 +38,7 @@ export function newlineBeforeBlockStatement(formattedText, ast) {
 		}
 
 		if (node.type === 'TryStatement') {
+
 			/*
 			 *                } catch (error) {
 			 *         block--^ ^--handler